Fix formatting types mismatch on Solaris

+++ Process.c
@@ -480,7 +480,9 @@ void Process_writeField(Process* this, RichString* str, ProcessField field) {
    case TIME: Process_printTime(str, this->time); return;
    case TGID: xSnprintf(buffer, n, Process_pidFormat, this->tgid); break;
    case TPGID: xSnprintf(buffer, n, Process_pidFormat, this->tpgid); break;
-   case TTY_NR: xSnprintf(buffer, n, "%3u:%3u ", major(this->tty_nr), minor(this->tty_nr)); break;
+   case TTY_NR:
+      xSnprintf(buffer, n, "%3u:%3u ", (unsigned int)major(this->tty_nr), (unsigned int)minor(this->tty_nr));
+      break;
    case USER: {
       if (Process_getuid != (int) this->st_uid)
          attr = CRT_colors[PROCESS_SHADOW];
